+++ tests/testthat/test_distance_funcs.R
@@ -93,4 +93,26 @@ test_that('compute_all_nearest_distance works', {
   expect_equal(all_d_2, nearby)
 
   file.remove(out_path)
+
+  # Repeat with a consolidated file
+  merge_path = file.path('test_data',
+                         'consolidated/Consolidated_data.txt')
+  out_path = tempfile()
+  compute_all_nearest_distance(merge_path, out_path)
+  expect_true(file.exists(out_path))
+  all_d = readr::read_tsv(out_path, na=c('NA', '#N/A'), col_types=readr::cols())
+  expect_equal(nrow(all_d), 66+68+270+215)
+
+  # The result for the sample data should be the same
+  csd = read_cell_seg_data(path)
+  nearby = find_nearest_distance(csd)
+  all_d_2 = all_d %>%
+    dplyr::filter(`Sample Name`=="FIHC4__0929309_HP_IM3_2.im3") %>%
+    dplyr::arrange(`Cell ID`) %>%
+    dplyr::select(`Distance to B`=`Distance to B+`,
+                  `Distance to Cytotoxic T`=`Distance to Cytotoxic_T+`,
+                  `Distance to Helper T`=`Distance to Helper_T+`)
+  expect_equal(all_d_2, nearby)
+
+  file.remove(out_path)
 })